third_party/WebKit/LayoutTests/platform/mac-snowleopard/http/tests/xmlhttprequest/origin-exact-matching-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Fwww2.localhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'http://www2.localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value '//localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value '://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=ftp%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'ftp://localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value 'http:://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'http:/localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3Alocalhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'http:localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=localhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%3F. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000?'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2F. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000/'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%20%2F. The 'Access-Control-Allow-Origin' header contains the invalid value 'http://localhost:8000 /'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%23. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000#'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2523. The 'Access-Control-Allow-Origin' header contains the invalid value 'http://localhost:8000%23'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%3A80. The 'Access-Control-Allow-Origin' header contains the invalid value 'http://localhost:8000:80'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2C%20*. The 'Access-Control-Allow-Origin' header contains the invalid value 'http://localhost:8000, *'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=HTTP%3A%2F%2FLOCALHOST%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'HTTP://LOCALHOST:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=HTTP%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' whitelists only 'HTTP://localhost:8000'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=-. The 'Access-Control-Allow-Origin' header contains the invalid value '-'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=**. The 'Access-Control-Allow-Origin' header contains the invalid value '**'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%00*. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%27*%27. The 'Access-Control-Allow-Origin' header contains the invalid value ''*''. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%22*%22. The 'Access-Control-Allow-Origin' header contains the invalid value '"*"'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*%20*. The 'Access-Control-Allow-Origin' header contains the invalid value '* *'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*http%3A%2F%2F*. The 'Access-Control-Allow-Origin' header contains the invalid value '*http://*'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*http%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value '*http://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*%20http%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value '* http://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*%2C%20http%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value '*, http://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%00http%3A%2F%2Flocalhost%3A8000. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=null%20http%3A%2F%2Flocalhost%3A8000. The 'Access-Control-Allow-Origin' header contains the invalid value 'null http://localhost:8000'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Fexample.net. The 'Access-Control-Allow-Origin' whitelists only 'http://example.net'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=null. The 'Access-Control-Allow-Origin' header contains the invalid value 'null'. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2Fxmlhttprequest%2Fresources%2Forigin-exact-matching-iframe.html. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000/xmlhttprequest/resources/origin-exact-matching-iframe.html'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2Fxmlhttprequest%2Fresources%2F. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000/xmlhttprequest/resources/'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=http%3A%2F%2Flocalhost%3A8000%2Fxmlhttprequest%2Fresources%2Forigin-exact-matching-iframe.html. The 'Access-Control-Allow-Origin' whitelists only 'http://localhost:8000/xmlhttprequest/resources/origin-exact-matching-iframe.html'. Origin 'http://localhost:8000' is not in the list, and is therefore not allowed access.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Check that exact matching is used when comparing a request's originating url and the value provided by Access-Control-Allow-Origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Should allow origin: '*'
PASS xhr.send(null) is undefined.
Should allow origin: ' *  '
PASS xhr.send(null) is undefined.
Should allow origin: '	*'
PASS xhr.send(null) is undefined.
Should allow origin: 'http://localhost:8000'
PASS xhr.send(null) is undefined.
Should allow origin: ' http://localhost:8000'
PASS xhr.send(null) is undefined.
Should allow origin: ' http://localhost:8000   	 '
PASS xhr.send(null) is undefined.
Should allow origin: '	http://localhost:8000'
PASS xhr.send(null) is undefined.
Should disallow origin: 'http://www2.localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '//localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'ftp://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http:://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http:/localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http:localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000?'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000/'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000 /'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000#'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000%23'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000:80'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000, *'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'HTTP://LOCALHOST:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'HTTP://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '-'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '**'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '*'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: ''*''
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '"*"'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '* *'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '*http://*'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '*http://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '* http://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: '*, http://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'null http://localhost:8000'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://example.net'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'null'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: ''
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000/xmlhttprequest/resources/origin-exact-matching-iframe.html'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000/xmlhttprequest/resources/'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
Should disallow origin: 'http://localhost:8000/xmlhttprequest/resources/origin-exact-matching-iframe.html'
PASS xhr.send(null) threw exception NetworkError: A network error occurred..
PASS successfullyParsed is true

TEST COMPLETE
